# swagger.json
{
...
  "swagger": "2.0",
  "securityDefinitions": {
  },
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "parameters": [
          {
            "in": "header",
            "default": "Bearer {token}",
            "name": "Authorization",
            "required": false,
            "type": "string"
          },
          {
            "schema": {
              "$ref": "#/definitions/POSTAuth-LoginRequest"
            },
            "in": "body",
            "name": "Body",
            "required": true,
            "description": ""
          }
        ],
        "summary": "POST Auth - Log in",
        "deprecated": false,
        "responses": {
          "200": {
            "headers": {
            },
            "description": ""
          }
        },
        "produces": [
          "application/json"
        ],
        "tags": [
          "Misc"
        ],
        "operationId": "POSTAuth-Login"
      }
    },
...
    "/transactions": {
      "get": {
        "parameters": [
          {
            "in": "header",
            "default": "Bearer {token}",
            "name": "Authorization",
            "required": false,
            "type": "string"
          }
        ],
        "summary": "GET Tranx - Get all infor",
        "deprecated": false,
        "responses": {
          "200": {
            "headers": {
            },
            "description": ""
          }
        },
        "produces": [
          "application/json"
        ],
        "tags": [
          "Misc"
        ],
        "operationId": "GETTranx-Getallinfor"
      },
      "post": {
        "parameters": [
          {
            "in": "header",
            "default": "Bearer {token}",
            "name": "Authorization",
            "required": false,
            "type": "string"
          },
          {
            "schema": {
              "$ref": "#/definitions/POSTTranx-CreatenewoneRequest"
            },
            "in": "body",
            "name": "Body",
            "required": true,
            "description": ""
          }
        ],
        "summary": "POST Tranx - Create new one",
        "deprecated": false,
        "responses": {
          "200": {
            "headers": {
            },
            "description": ""
          }
        },
        "produces": [
          "application/json"
        ],
        "tags": [
          "Misc"
        ],
        "operationId": "POSTTranx-Createnewone"
      }
    },
...
  },
  "host": "localhost:8080",
  "consumes": [
    "application/json"
  ],
  "basePath": "/api",
  "produces": [
    "application/json"
  ],
  "definitions": {
    "POSTTranx-CreatenewoneRequest": {
      "properties": {
        "bookId": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        }
      },
      "title": "POSTTranx-CreatenewoneRequest",
      "type": "object",
      "example": {
        "bookId": "5f1fcd0cdb5f44abea1b74cb",
        "userId": "5f1ffc75e9d40a02847321f1"
      },
      "required": [
        "userId",
        "bookId"
      ]
    },
...
    "POSTAuth-LoginRequest": {
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "title": "POSTAuth-LoginRequest",
      "type": "object",
      "example": {
        "email": "longnt8910@gmail.com",
        "password": "123456"
      },
      "required": [
        "email",
        "password"
      ]
    },
...
  },
  "info": {
    "contact": {
    },
    "title": "Express-27-02-API",
    "description": "API list for Express-27-02",
    "version": "1.0"
  }
}